README.md
Human Detection and SMS Alert System:
This project utilizes OpenCV and cvzone's PoseModule to detect human presence in a video feed. If a human is continuously detected, an SMS alert is sent using Twilio. Ideal for security and surveillance applications, the system ensures real-time monitoring and automated alerting for enhanced safety. 🚀